Fix issue with updates
