+++ src/app/shared/components/data-display/data-display.component.ts
@@ -33,7 +33,7 @@ export class DataDisplayComponent implements OnInit {
       cell: (row: Visit) => formatDuration(row.startedAt, row.endedAt, row.timezone),
     },
   ];
-  loading = true;
+  loading: boolean = true;
 
   constructor() {}
