+++ src/app/teams/components/selected-team-toolbar/selected-team-toolbar.component.scss
@@ -4,7 +4,7 @@
   flex-wrap: nowrap;
   background-color: #f5f5f5;
   align-items: center;
-  height: 63px; // Leave 1px for divider
+  height: 64px;
 
   .selected-team-toolbar__header {
     font: 500 20px/32px Roboto,"Helvetica Neue",sans-serif;
